readStocks reducer created with thunk
